+++ config/handlers.php
@@ -19,6 +19,8 @@ return [
     | Here you may specify which class will be used as the base class
     | for all generated handlers.
     |
+    | Default is `\Laravel\Handlers\Handler::class` , but it only work with Laravel.
+    | Alternative value is `\App\Http\Controllers\Controller::class` (work with both Laravel and Lumen)
     */
     'base' => \Laravel\Handlers\Handler::class,
 ];