+++ config/routes.php
@@ -1,5 +1,0 @@
-<?php
-
-use Illuminate\Support\Facades\Route;
-
-//Route::post('/login', \App\Http\Handlers\LoginHandler::class);