+++ tests/Unit/HandlerTest.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Laravel\Handlers\Tests\Unit;
+
+use BadMethodCallException;
+use Laravel\Handlers\Handler;
+use Laravel\Handlers\Tests\TestCase;
+
+/**
+ * Class HandlerTest
+ *
+ * @package     Laravel\Handlers\Tests\Unit
+ * @author      Oanh Nguyen <oanhnn.bk@gmail.com>
+ * @license     The MIT License
+ */
+class HandlerTest extends TestCase
+{
+    /**
+     *
+     */
+    public function testCallInvokeMethod()
+    {
+        $this->assertTrue(true);
+    }
+}