+++ CHANGELOG.md
@@ -26,9 +26,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [0.1.2] - 2018-10-08
+
+### Removed
+- Remove register routes in provider
+
+
+
+## [0.1.1] - 2018-10-08
+
+### Changed
+- Update stub file and TravisCI config
+
+
+
 ## 0.1.0 - 2018-10-08
 
 First release
 
 
-[Unreleased]: https://github.com/oanhnn/laravel-handlers/compare/v0.1.0...develop
+
+[Unreleased]: https://github.com/oanhnn/laravel-handlers/compare/0.1.2...develop
+[0.1.2]: https://github.com/oanhnn/laravel-handlers/compare/0.1.1...0.1.2
+[0.1.1]: https://github.com/oanhnn/laravel-handlers/compare/0.1.0...0.1.1